Changed the args for FTContour::AddPoint() to float to get rid of warning. Also make more sense.

+++ include/FTVectoriser.h
@@ -45,7 +45,7 @@ class FTContour
 		FTContour();
 		~FTContour();
 	
-		void AddPoint( const int x, const int y);
+		void AddPoint( const float x, const float y);
 		
 		int size() const { return pointList.size();}
